Vue container as 2nd implementation of DI interface
